requirements.txt
sleekxmpp
dnspython
email
systemd-python
py-zabbix
